requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
-e git+https://github.com/haycacelik/AutoML-Final_Project.git@216f7db0cb7693152aba371ac187f9608ac36949#egg=automl
botorch==0.12.0
certifi==2025.7.14
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
ConfigSpace==1.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
gpytorch==1.13
greenlet==3.2.3
grpcio==1.73.1
hf-xet==1.1.5
hpbandster==0.7.4
huggingface-hub==0.33.4
idna==3.10
ifBO==0.3.11
ipython==8.37.0
jaxtyping==0.2.19
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kaleido==1.0.0
kiwisolver==1.4.8
linear-operator==0.5.3
logistro==1.1.0
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multipledispatch==1.0.0
multiprocess==0.70.16
narwhals==1.48.1
netifaces==0.11.0
networkx==2.8.8
neural-pipeline-search==0.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opt_einsum==3.4.0
optuna==4.4.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
portalocker==2.10.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyro-api==0.1.2
pyro-ppl==1.9.1
Pyro4==4.82
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.48.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.33.2
serpent==1.41
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.5
submitit==1.5.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
triton==3.3.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
